Rename full_account.hpp to api_objects.hpp
